+++ src/core/AbstractTitrationModel.cpp
@@ -44,6 +44,7 @@ AbstractTitrationModel::AbstractTitrationModel(DataClass* data)
     IndependentModel()->setHeaderData(0, Qt::Horizontal, "Host (A)", Qt::DisplayRole);
     IndependentModel()->setHeaderData(1, Qt::Horizontal, "Guest (B)", Qt::DisplayRole);
     m_ylabel = "&delta; [ppm]";
+    LoadSystemParameter();
 }
 
 AbstractTitrationModel::AbstractTitrationModel(AbstractTitrationModel* other)
@@ -52,10 +53,22 @@ AbstractTitrationModel::AbstractTitrationModel(AbstractTitrationModel* other)
     IndependentModel()->setHeaderData(0, Qt::Horizontal, "Host (A)", Qt::DisplayRole);
     IndependentModel()->setHeaderData(1, Qt::Horizontal, "Guest (B)", Qt::DisplayRole);
     m_ylabel = "&delta; [ppm]";
+    m_T = other->m_T;
 }
 
 AbstractTitrationModel::~AbstractTitrationModel()
 {
+}
+
+void AbstractTitrationModel::DeclareSystemParameter()
+{
+    addSystemParameter(Temperature, "Temperature", "Temperature in K", SystemParameter::Scalar);
+    setSystemParameterValue(Temperature, 298);
+}
+
+void AbstractTitrationModel::UpdateParameter()
+{
+    m_T = getSystemParameter(Temperature).Double();
 }
 
 void AbstractTitrationModel::DeclareOptions()