+++ src/core/models/itc/itc_1_1_Model.h
@@ -89,7 +89,7 @@ public:
             return i + 3;
     }
 
-    virtual QString AdditionOutput() const override;
+    virtual QString AdditionalOutput() const override;
 
 protected:
     virtual void CalculateVariables() override;