+++ src/core/models/titrations/1_1_Model.h
@@ -69,6 +69,7 @@ public:
         else
             return i + 3;
     }
+    virtual QString AdditionalOutput() const override;
 
 protected:
     virtual void CalculateVariables() override;