+++ src/ui/widgets/statisticwidget.cpp
@@ -76,6 +76,7 @@ void StatisticWidget::Update()
     overview += "<tr><td>cov<sub>fit</sub>:</td><td><b>" + Print::printDouble(m_model.data()->CovFit()) + "</b></td></tr>\n";
 
     overview += "</table>";
+    overview += m_model.data()->AdditionalOutput();
 
     m_short = overview;
     overview.clear();