+++ Changelog.md
@@ -1,5 +1,7 @@
 # SupraFit 2.0 pre-Alpha
 
+- Improved thermodynamic output -> calculation of free enthalpy and entropy (if heat available) including confidence
+- ToolSet::Confidence functions gives identical results compared to the octave quantile function
 - More consistent parameter names 
 - MetaModels are uniquely created and SupraFit doesn't crash due to wrongly
 override system parameters on drag and drop in the project list